Renamed object builders to strategies
